pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sidekick"
version = "0.1.0"
description = "Personal AI Agent with pydantic-ai"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "pydantic-ai-slim[google,logfire,mcp]>=1.0.0",
    "google-genai>=1.0.0",
    "litellm>=1.0.0",
    "slack-bolt>=1.27.0",
    "aiohttp>=3.9.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "exa-py>=1.0.0",
    "apscheduler>=3.10.0",
    "sqlalchemy>=2.0.0",
    "slowapi>=0.1.9",
    "mlx-audio>=0.3.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module level import not at top (load_dotenv pattern)
    "B007",   # loop control variable not used (intentional in some cases)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "ARG001", # unused function argument (common in tools)
    "ARG002", # unused method argument (common in tests for fixtures)
    "SIM103", # return condition directly (sometimes less readable)
    "SIM105", # contextlib.suppress (sometimes less clear)
    "SIM110", # use all() instead of for loop (sometimes less readable)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F841"]  # unused variables OK in tests (side effects)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
